revert json-serializable rename
