+++ shared/json-serializable.ts
